Remove shortcut, but also avoid some allocations
Day 8 was also more challenging for me, because I additionally tried to
work with very few allocations if possible.
Passing around iterators is weird at first, but this day is a good
example of how to achieve it.
It even carries nested iterators! Noice!!!

Useful article:
https://depth-first.com/articles/2020/06/22/returning-rust-iterators/

(This is maybe the pixie dust and magic, a certain someone metioned to
me. I like it though. ;-))