Refactor day 12
- use &str instead of String for Id
- remove enum and work on &str's instead (for simplicity reasons)
- - the enum was a nice idea, but didn't help that much tbh
- use references everywhere if possible
- thread through all lifetime bounds - what a weaver of Saidin I am!

This has the advantage that we basically work with the single set of
input str(ings) and don't allocate new Strings for both hashmaps.
Also the clone of the visits hashmap should still use the same
references for the keys (I hope).
So the only allocations left are visits cloning, which are still plenty.
And I guess are necessary evil, as we cannot know in advance how much we
will need for finding all paths.

Wins after all the changes from initial solution:
- debug mode runtime from 5s down to 1.5s (3.3 times faster)
- release mode runtime from 337ms down to 56ms (6 times faster)