Improve performance even further
Collected some inspiration from the AoC subreddit.
- switch (back) to a Node enum, but with u32 as data for Small/Big
- use ahash instead of std's hashmap
- use hash32 for node data creation

Not all things add a lot of improvements, the hash32/u32 was an attempt
to reduce the size of the enum (with u64:16 bytes, u32 only 8 bytes),
but it gains only around 1ms max (probably less).
Also ahash was not a very huge deal on its own.
But the combination of all the things got me 11ms over previous version.

Performance results:
- debug mode: 5s -> 514ms (9.7 times faster)
- release mode: 337ms -> 28ms (12 times faster)